pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envsmith"
version = "0.1.0"
description = "Modern, schema-first environment variable loader and manager for Python."
authors = [
    { name = "Peyman Tahmasebi", email = "peymantahmasebi10@gmail.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100.0"]
django = ["django>=4.0.0"]

[project.urls]
Homepage = "https://github.com/PeymanSohi/envsmith"
Documentation = "https://github.com/PeymanSohi/envsmith#readme"
Source = "https://github.com/PeymanSohi/envsmith"
Changelog = "https://github.com/PeymanSohi/envsmith/blob/main/CHANGELOG.md"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["envsmith"]
